Add module github_users.
