+++ github_users/src/main/java/com/wiredcraft/github_users/bean/DataResponseBean.kt
