+++ github_users/src/main/java/com/wiredcraft/github_users/util/NetworkUtil.kt
