+++ github_users/src/main/res/drawable-v21/ripple_drawable.xml
@@ -1,0 +1,8 @@
+<ripple xmlns:android="http://schemas.android.com/apk/res/android"
+    android:color="#bdbdbd">
+    <item android:id="@android:id/mask">
+        <shape android:shape="rectangle">
+            <solid android:color="#f0f0f0" />
+        </shape>
+    </item>
+</ripple>
