+++ github_users/src/main/res/drawable/ripple_drawable.xml
@@ -1,0 +1,12 @@
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_pressed="true">
+        <shape android:shape="rectangle">
+            <solid android:color="#bdbdbd" />
+        </shape>
+    </item>
+    <item android:state_pressed="false">
+        <shape android:shape="rectangle">
+            <solid android:color="#f0f0f0" />
+        </shape>
+    </item>
+</selector>
