+++ github_users/src/main/res/layout/activity_user_detail.xml
