+++ github_users/src/main/res/layout/layout_footer.xml
