+++ github_users/src/main/res/layout/layout_title_bar.xml
